+++ R/classes_methods.R
@@ -2065,15 +2065,8 @@ MergeRareHaplotypes.RADdata <- function(object, min.ind.with.haplotype = 10,
     while(length(theserare) > 0 && length(thesealleles) > 1){
       thisAl <- theserare[1]
       # get nucleotide distance between this allele and others
-      thesenuc <- object$alleleNucleotides[thesealleles]
-      splitnuc <- strsplit(thesenuc, split = "")
-      splitnucA <- splitnuc[[which(thesealleles == thisAl)]]
-      nucdist <- integer(length(thesealleles))
-      for(a in 1:length(thesealleles)){
-        nucdist[a] <- sum(sapply(1:length(splitnucA),
-                                 function(i) polyRADsubmat[splitnucA[i],
-                                                           splitnuc[[a]][i]]))
-      }
+      nucdist <- .nucdist(object$alleleNucleotides[thisAl],
+                          object$alleleNucleotides[thesealleles])
       # find the closest allele
       alToMerge <- thesealleles[nucdist == min(nucdist[-match(thisAl, thesealleles)])]
       if(length(alToMerge) > 1){
@@ -2443,33 +2436,39 @@ MergeIdenticalHaplotypes.RADdata <- function(object, ...){
   if(!is.null(object$alleleFreq) || !is.null(object$depthSamplingPermutations)){
     stop("Run MergeIdenticalHaplotypes before running any pipeline functions.")
   }
+  varsiteonly <- attr(object$alleleNucleotides, "Variable_sites_only")
   
   remal <- integer(0) # indices of alleles to remove
   for(L in 1:nLoci(object)){
     thesecol <- which(object$alleles2loc == L)
-    dup <- duplicated(object$alleleNucleotides[thesecol])
-    for(al in thesecol[dup]){
-      # find allele to merge this one into
-      alM <- min(thesecol[object$alleleNucleotides[thesecol] == 
-                            object$alleleNucleotides[al]])
-      stopifnot(al != alM)
-      # consolidate read depth
-      object$alleleDepth[,alM] <- object$alleleDepth[,alM] +
-        object$alleleDepth[,al]
-      object$antiAlleleDepth[,alM] <- object$antiAlleleDepth[,alM] -
-        object$alleleDepth[,al]
+    nucident <- .nucdist(object$alleleNucleotides[thesecol]) == 0
+    diag(nucident) <- FALSE # don't need to merge allele to self
+    while(any(nucident)){
+      al <- which(rowSums(nucident) > 0)[1] # merge alleles into this one
+      alM <- which(nucident[al,]) # alleles to merge
+      object$alleleDepth[,thesecol[al]] <- object$alleleDepth[,thesecol[al]] +
+        rowSums(object$alleleDepth[,thesecol[alM], drop = FALSE])
+      object$antiAlleleDepth[,thesecol[al]] <- object$antiAlleleDepth[,thesecol[al]] +
+        rowSums(object$antiAlleleDepth[,thesecol[alM], drop = FALSE])
+      for(am in alM){
+        object$alleleNucleotides[thesecol[al]] <-
+          .mergeNucleotides(object$alleleNucleotides[thesecol[al]],
+                            object$alleleNucleotides[thesecol[am]])
+      }
+      remal <- c(remal, thesecol[alM])
+      nucident <- nucident[-alM, -alM]
     }
-    remal <- c(remal, thesecol[dup])
   }
   
   # remove duplicated alleles from all slots
-  object$alleleDepth <- object$alleleDepth[,-remal]
-  object$antiAlleleDepth <- object$antiAlleleDepth[,-remal]
-  object$alleles2loc <- object$alleles2loc[-remal]
-  varsiteonly <- attr(object$alleleNucleotides, "Variable_sites_only")
-  object$alleleNucleotides <- object$alleleNucleotides[-remal]
+  if(length(remal) > 0){
+    object$alleleDepth <- object$alleleDepth[,-remal]
+    object$antiAlleleDepth <- object$antiAlleleDepth[,-remal]
+    object$alleles2loc <- object$alleles2loc[-remal]
+    object$alleleNucleotides <- object$alleleNucleotides[-remal]
+    object$depthRatio <- object$depthRatio[,-remal]
+  }
   attr(object$alleleNucleotides, "Variable_sites_only") <- varsiteonly
-  object$depthRatio <- object$depthRatio[,-remal]
   
   return(object)
 }